+++ ubo_app/utils/server.py
@@ -11,17 +11,15 @@ from ubo_app.utils import IS_RPI
 
 
 @overload
-async def send_command(command: str) -> None: ...
-
-
+async def send_command(*command: str) -> None: ...
 @overload
-async def send_command(command: str, *, has_output: Literal[True]) -> str | None: ...
-
-
-async def send_command(command: str, *, has_output: bool = False) -> str | None:
+async def send_command(*command: str, has_output: Literal[True]) -> str | None: ...
+async def send_command(*command_: str, has_output: bool = False) -> str | None:
     """Send a command to the system manager socket."""
     if not IS_RPI:
         return None
+
+    command = ' '.join(command_)
 
     try:
         reader, writer = await asyncio.open_unix_connection(SERVER_SOCKET_PATH)
@@ -39,6 +37,6 @@ async def send_command(command: str, *, has_output: bool = False) -> str | None:
 
     except Exception:
         logger.exception('Failed to send command to the server')
-        return None
+        raise
     else:
         return response